refactor: drop redundant Float64 conversion in gaussian_bin_scale
- 2.0 and df (from frequency_bin_width) are already Float64, so Julia
  promotion yields Float64 for Int/Float32/Rational observation times
  with identical rounding
- the production caller already passes Float64 via year_to_second

+++ AstroSGWB/src/detector/observation.jl
@@ -19,5 +19,5 @@ function gaussian_bin_scale(;
 )
     df = frequency_bin_width(frequencies)
     # effective_psd is amplitude √(variance); bin variance is (effective_psd)² / (2 T Δf)
-    return sqrt.(effective_psd .^ 2 ./ (2.0 * Float64(observation_time_sec) * df))
+    return sqrt.(effective_psd .^ 2 ./ (2.0 * observation_time_sec * df))
 end
